Renamed module types to type Fix: #362
