+++ src/electionguard/key_ceremony_mediator.py
@@ -11,7 +11,7 @@ from .key_ceremony import (
     combine_election_public_keys,
     verify_election_partial_key_challenge,
 )
-from .types import GUARDIAN_ID, MEDIATOR_ID
+from .type import GUARDIAN_ID, MEDIATOR_ID
 
 
 class GuardianPair(NamedTuple):
