+++ tests/integration/test_functional_key_ceremony.py
@@ -15,7 +15,7 @@ from electionguard.key_ceremony import (
     verify_election_partial_key_backup,
     verify_election_partial_key_challenge,
 )
-from electionguard.types import GUARDIAN_ID
+from electionguard.type import GUARDIAN_ID
 from electionguardtest.identity_encrypt import identity_auxiliary_encrypt
