+++ tests/unit/test_decryption.py
@@ -37,7 +37,7 @@ from electionguard.guardian import Guardian
 from electionguard.key_ceremony import CeremonyDetails, ElectionKeyPair
 from electionguard.key_ceremony_mediator import KeyCeremonyMediator
 from electionguard.tally import tally_ballots
-from electionguard.types import BALLOT_ID, GUARDIAN_ID
+from electionguard.type import BALLOT_ID, GUARDIAN_ID
 from electionguard.utils import get_optional
 
 import electionguardtest.ballot_factory as BallotFactory
